requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
aiosignal==1.2.0
arrow==1.2.2
asgiref==3.5.0
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
bidict==0.21.4
binaryornot==0.4.4
botbuilder-core==4.14.2
botbuilder-integration-aiohttp==4.14.2
botbuilder-schema==4.14.2
botframework-connector==4.14.2
botframework-streaming==4.14.2
boto3==1.22.11
botocore==1.25.11
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cookiecutter==1.7.0
cryptography==3.4.8
cycler==0.11.0
decorator==5.1.1
deepgram-sdk==0.2.1
distlib==0.3.4
dnspython==2.2.0
EZGmail==2022.2.24
fasteners==0.17.2
filelock==3.4.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
flatbuffers==2.0
fonttools==4.28.5
frozenlist==1.2.0
future==0.18.2
gast==0.4.0
gevent==21.12.0
gevent-websocket==0.10.1
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
greenlet==1.1.2
grpcio==1.43.0
h5py==3.6.0
httplib2==0.20.4
idna==2.10
importlib-metadata==4.10.0
ipython==7.31.0
isodate==0.6.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==1.0.0
joblib==1.1.0
jsonpickle==1.4.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
msal==1.6.0
msrest==0.6.19
multidict==5.2.0
nltk==3.6.7
numpy==1.21.5
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
poyo==0.5.0
prompt-toolkit==3.0.24
protobuf==3.19.3
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyJWT==1.7.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.1
pytz==2021.3
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tesserocr @ file:///C:/Users/kavin/Downloads/tesserocr-2.4.0-cp37-cp37m-win_amd64.whl
threadpoolctl==3.0.0
tqdm==4.62.3
traitlets==5.1.1
twilio==7.5.0
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.0
wcwidth==0.2.5
websockets==10.1
Werkzeug==2.0.2
wfastcgi==3.0.0
whichcraft==0.6.1
wrapt==1.13.3
yarl==1.4.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
